tests/wpt/web-platform-tests/speculation-rules/prerender/cookies.html

<!DOCTYPE html>
<title>Same-origin prerendering can access cookies</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-channel');

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  const initiator_cookie = 'initiator_cookie=exist';
  const prerender_cookie = 'prerender_cookie=exist';

  document.cookie = initiator_cookie;

  // Start prerendering a page that attempts to access cookies.
  startPrerendering(`resources/cookies-access.html`);
  const result = await gotMessage;
  assert_equals(
    result, initiator_cookie,
    'prerendering page should be able to read from document cookies.');
  assert_equals(
    document.cookie, initiator_cookie + '; ' + prerender_cookie,
    'prerendering page should be able to write to document cookies');
}, 'prerendering page should be able to access cookies');

</script>
</body>
